+++ examples/README.md
@@ -1,0 +1,8 @@
+## Example plugins
+
+This folder contains example plugins.  To run the plugins in this folder, use the `--run-examples` flag, via
+
+```
+yarn start --run-examples
+```
+
