+++ examples/demo_search/kibana.json
