+++ examples/demo_search/package.json
@@ -8,7 +8,7 @@
   },
   "license": "Apache-2.0",
   "scripts": {
-    "kbn": "node ../../../../scripts/kbn.js",
+    "kbn": "node ../../scripts/kbn.js",
     "build": "rm -rf './target' && tsc"
   },
   "devDependencies": {
