+++ examples/demo_search/public/index.ts
