+++ examples/demo_search/server/index.ts
