+++ examples/search_explorer/kibana.json
