+++ examples/search_explorer/public/documentation.tsx
