+++ examples/search_explorer/public/guide_section.tsx
