+++ examples/search_explorer/public/index.ts
