+++ examples/search_explorer/public/page.tsx
