+++ examples/search_explorer/tsconfig.json
@@ -1,16 +1,15 @@
 {
-  "extends": "../../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "./target",
     "skipLibCheck": true
   },
   "include": [
     "index.ts",
-    "common/**/*.ts",
     "public/**/*.ts",
     "public/**/*.tsx",
     "server/**/*.ts",
-    "../../../../typings/**/*"
+    "../../typings/**/*",
   ],
   "exclude": []
 }
