+++ scripts/functional_tests.js
@@ -24,4 +24,5 @@ require('@kbn/test').runTestsCli([
   require.resolve('../test/plugin_functional/config.js'),
   require.resolve('../test/interpreter_functional/config.ts'),
   require.resolve('../test/ui_capabilities/newsfeed_err/config.ts'),
+  require.resolve('../test/examples/config.js')
 ]);
