+++ test/examples/search/index.ts
